Revert "Enable PRELINK" and update prelink map
This reverts commit a1d2e88ca8114486ba8adb29bc7f0e810261f638 because PRELINK
is not actually enabled by simply removing "TARGET_PRELINK_MODULE := false".

This leaves PRELINK disabled for now, although the bug has been fixed in lp:825327,
because no obvious performance improvement is observed when it is enabled. The
comment in core/config.mk is updated accordingly.

There is an error when PRELINK is enabled, which is due to missing libnl.so in
prelink-linux-arm.map. This error is fixed by allocating memory for libnl.so
in the prelink map file.

Change-Id: Id4d2c39f704988b7e62413494ad81da3ec0e501f